Refactor EmailFlowStart
