+++ front/app/containers/Authentication/steps/EmailFlowStart/index.tsx
@@ -1,9 +1,6 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 
-import { Box, Text } from '@citizenlab/cl2-component-library';
-import { yupResolver } from '@hookform/resolvers/yup';
-import { useForm, FormProvider } from 'react-hook-form';
-import { string, object } from 'yup';
+import { Box } from '@citizenlab/cl2-component-library';
 
 import { SSOProvider } from 'api/authentication/singleSignOn';
 
@@ -11,20 +8,12 @@ import oldMessages from 'containers/Authentication/steps/_components/AuthProvide
 import { SetError } from 'containers/Authentication/typings';
 import useAuthConfig from 'containers/Authentication/useAuthConfig';
 
-import Input from 'components/HookForm/Input';
-import ButtonWithLink from 'components/UI/ButtonWithLink';
 import FranceConnectButton from 'components/UI/FranceConnectButton';
 import Or from 'components/UI/Or';
 
 import { useIntl } from 'utils/cl-intl';
-import {
-  isCLErrorsWrapper,
-  handleHookFormSubmissionError,
-} from 'utils/errorUtils';
-import { isValidEmail } from 'utils/validate';
 
-import sharedMessages from '../messages';
-
+import EmailForm from './EmailForm';
 import SSOButtons from './SSOButtons';
 
 interface Props {
@@ -35,14 +24,6 @@ interface Props {
   onEnterFranceConnect: () => void;
 }
 
-interface FormValues {
-  email: string;
-}
-
-const DEFAULT_VALUES: Partial<FormValues> = {
-  email: undefined,
-};
-
 const EmailFlowStart = ({
   loading,
   setError,
@@ -51,42 +32,7 @@ const EmailFlowStart = ({
   onEnterFranceConnect,
 }: Props) => {
   const { passwordLoginEnabled, ssoProviders } = useAuthConfig();
-
   const { formatMessage } = useIntl();
-
-  const schema = useMemo(
-    () =>
-      object({
-        email: string()
-          .email(formatMessage(sharedMessages.emailFormatError))
-          .required(formatMessage(sharedMessages.emailMissingError))
-          .test(
-            '',
-            formatMessage(sharedMessages.emailFormatError),
-            isValidEmail
-          ),
-      }),
-    [formatMessage]
-  );
-
-  const methods = useForm<FormValues>({
-    mode: 'onSubmit',
-    defaultValues: DEFAULT_VALUES,
-    resolver: yupResolver(schema),
-  });
-
-  const handleSubmit = async ({ email }: FormValues) => {
-    try {
-      await onSubmit(email);
-    } catch (e) {
-      if (isCLErrorsWrapper(e)) {
-        handleHookFormSubmissionError(e, methods.setError);
-        return;
-      }
-
-      setError('unknown');
-    }
-  };
 
   return (
     <Box data-cy="email-flow-start">
@@ -106,32 +52,7 @@ const EmailFlowStart = ({
         </>
       )}
       {passwordLoginEnabled && (
-        <FormProvider {...methods}>
-          <form onSubmit={methods.handleSubmit(handleSubmit)}>
-            <Text mt="0px" mb="32px" color="tenantText">
-              {formatMessage(sharedMessages.enterYourEmailAddress)}
-            </Text>
-            <Box data-cy="email-flow-start-email-input">
-              <Input
-                name="email"
-                type="email"
-                autocomplete="email"
-                label={formatMessage(sharedMessages.email)}
-              />
-            </Box>
-            <Box w="100%" display="flex" mt="32px">
-              <ButtonWithLink
-                dataCy="email-flow-start-continue-button"
-                type="submit"
-                width="100%"
-                disabled={loading}
-                processing={loading}
-              >
-                {formatMessage(sharedMessages.continue)}
-              </ButtonWithLink>
-            </Box>
-          </form>
-        </FormProvider>
+        <EmailForm loading={loading} setError={setError} onSubmit={onSubmit} />
       )}
       <SSOButtons onClickSSO={onSwitchToSSO} />
     </Box>
