+++ front/app/containers/Authentication/steps/EmailFlowStart/types.ts
@@ -1,0 +1,3 @@
+export interface FormValues {
+  email: string;
+}
